+++ packages/sampler/CHANGELOG.md
@@ -2,13 +2,13 @@
 
 The following is a curated list of changes in the Enact Sampler, newest changes on the top.
 
-## [2.0.0-beta.6] - 2018-06-04
+## [2.0.0-beta.7] - 2018-06-11
 
 No significant changes.
 
-### Fixed
+## [2.0.0-beta.6] - 2018-06-04
 
-- `Slider` in QA sampler to deprecate `data` prop in `moonstone/VirtualList`
+No significant changes.
 
 ## [2.0.0-beta.5] - 2018-05-29